requirements.txt
Django==1.5.12
python-dateutil==2.4.2
python-mimeparse==0.1.4
django-tastypie==0.12.2
six==1.9.0
South==0.7.6